pyproject.toml
[project]
name = "swarmpal-test-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
    "swarmpal",
    "xarray>=2025.1.2",
]

[tool.uv.sources]
swarmpal = { git = "https://github.com/Swarm-DISC/SwarmPAL", branch = "staging" }
